requirements.txt
langchain
langchain-community
langchain_groq
chromadb
streamlit
pandas
python-dotenv
langchain_huggingface
sentence-transformers
streamlit